Update README.md
Credit card fraud detection models are typically trained on a dataset of historical credit card transaction data. This data includes information about the transactions, such as the amount of the transaction, the merchant where the transaction was made, and the customer who made the transaction. The model is trained to identify transactions that are likely to be fraudulent.

Once the model is trained, it can be used to detect fraudulent transactions in real time. When a new transaction is made, the model is used to predict whether or not the transaction is fraudulent. If the transaction is predicted to be fraudulent, it can be blocked.

Credit card fraud detection models can help businesses to reduce their risk of fraud. By identifying fraudulent transactions in real time, businesses can prevent these transactions from being processed and can avoid financial losses